+++ CHANGELOG.md
@@ -10,6 +10,7 @@ All notable changes qsf-integration project will be documented in this file.
 - make it possible to use string and date in range queries for elastic
 - implement a search filter builder
 - implement a not filter for the qsfql query
+- add the request id to the pipeline container
 
 ### Changed